+++ custom_components/meraki_ha/core/api/endpoints/appliance/vpn.py
@@ -6,7 +6,7 @@ import logging
 from typing import TYPE_CHECKING, Any
 
 from custom_components.meraki_ha.core.api.cache import async_timed_cache
-from custom_components.meraki_ha.core.utils.api_utils import (
+from custom_components.meraki_ha.core.utils.api import (
     handle_meraki_errors,
     validate_response,
 )
